+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### 🔧 Fixed
+
+* `hal auto --modify --count X` was not refreshing the `ExposureHelper` after updating the parameters.
+
+
 ## 0.6.0b1 - January 10, 2023
 
 ### 🚀 New